+++ FRONTEND/asset/icon/site.webmanifest
@@ -3,12 +3,12 @@
     "short_name": "Portfolio Tracker",
     "icons": [
         {
-            "src": "asset/icon/android-chrome-192x192.png",
+            "src": "android-chrome-192x192.png",
             "sizes": "192x192",
             "type": "image/png"
         },
         {
-            "src": "asset/icon/android-chrome-512x512.png",
+            "src": "android-chrome-512x512.png",
             "sizes": "512x512",
             "type": "image/png"
         }
