+++ FRONTEND/dividend.html
@@ -79,8 +79,6 @@
                                     Dividend</a
                                 >
                             </li>
-
-                            <!-- <li class="nav-item dropdown"><a id="dropdownSubMenu1" href="#" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" class="nav-link dropdown-toggle" >Dropdown</a ><ul aria-labelledby="dropdownSubMenu1" class="dropdown-menu border-0 shadow" ><li><a href="#" class="dropdown-item" >Some action</a></li><li><a href="#" class="dropdown-item" >Some other action</a ></li><li class="dropdown-divider"></li><li class="dropdown-submenu dropdown-hover"><a id="dropdownSubMenu2" href="#" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" class="dropdown-item dropdown-toggle" >Hover for action</a ><ul aria-labelledby="dropdownSubMenu2" class="dropdown-menu border-0 shadow" ><li><a tabindex="-1" href="#" class="dropdown-item" >level 2</a ></li><li class="dropdown-submenu"><a id="dropdownSubMenu3" href="#" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" class="dropdown-item dropdown-toggle" >level 2</a ><ul aria-labelledby="dropdownSubMenu3" class="dropdown-menu border-0 shadow" ><li><a href="#" class="dropdown-item" >3rd level</a ></li><li><a href="#" class="dropdown-item" >3rd level</a ></li></ul></li><li><a href="#" class="dropdown-item" >level 2</a></li><li><a href="#" class="dropdown-item">level 2</a></li></ul></li></ul></li> -->
                         </ul>
                     </div>
                 </div>
@@ -90,12 +88,15 @@
                     <div class="container">
                         <div class="row mb-2">
                             <div class="col-sm-6">
-                                <h1 class="m-0">Home</h1>
+                                <h1 class="m-0">Dividend</h1>
                             </div>
                             <div class="col-sm-6">
                                 <ol class="breadcrumb float-sm-right">
-                                    <li class="breadcrumb-item active">
+                                    <li class="breadcrumb-item">
                                         <a href="#">Home</a>
+                                    </li>
+                                    <li class="breadcrumb-item active">
+                                        Dividend
                                     </li>
                                 </ol>
                             </div>
@@ -104,43 +105,58 @@
                 </div>
                 <div class="content">
                     <div class="container">
-                        <div class="row" id="FinancialSummary"></div>
                         <div class="row">
-                            <div class="col">
+                            <div class="col-lg-6">
                                 <div class="card">
                                     <div class="card-header">
-                                        <div
-                                            class="d-flex justify-content-between"
+                                        <h3 class="card-title">
+                                            Dividend Info
+                                        </h3>
+                                    </div>
+                                    <div class="card-body p-0">
+                                        <div class="table-responsive">
+                                            <table
+                                                id="DividendTable"
+                                                class="table m-0 table-bordered table-striped"
+                                            ></table>
+                                        </div>
+                                    </div>
+                                    <div class="card-footer clearfix">
+                                        <a
+                                            href="javascript:void(0)"
+                                            class="btn btn-sm btn-secondary float-right"
+                                            >View All Orders</a
                                         >
-                                            <h3 class="card-title">
-                                                Portfolio Performance
-                                            </h3>
-                                            <a href="">View Report</a>
+                                    </div>
+                                </div>
+                            </div>
+
+                            <div class="col-lg-6">
+                                <div class="card">
+                                    <div class="card-header">
+                                        <div class="d-flex">
+                                            <div class="d-flex flex-column">
+                                                <h3 class="card-title">
+                                                    Yearly Dividend Info
+                                                </h3>
+                                            </div>
+                                            <p
+                                                class="ml-auto d-flex flex-column text-right"
+                                            >
+                                                <span
+                                                    class="text-success text-bold text-lg"
+                                                    id="totalDividend"
+                                                    >$18,230.00</span
+                                                >
+                                                <span class="text-muted"
+                                                    >Total Dividend</span
+                                                >
+                                            </p>
                                         </div>
                                     </div>
                                     <div class="card-body">
                                         <div class="position-relative mb-4">
-                                            <div
-                                                class="chart"
-                                                id="holdingsTrandsChart"
-                                                style="height: 20rem"
-                                            ></div>
-                                        </div>
-                                        <div
-                                            class="d-flex flex-row justify-content-end"
-                                        >
-                                            <span class="mr-2">
-                                                <i
-                                                    class="fas fa-square text-primary"
-                                                ></i>
-                                                Investment
-                                            </span>
-                                            <span>
-                                                <i
-                                                    class="fas fa-square text-success"
-                                                ></i>
-                                                Market Value
-                                            </span>
+                                            <div id="dividendBarChart"></div>
                                         </div>
                                     </div>
                                 </div>
@@ -162,6 +178,6 @@
                 </div>
             </footer>
         </div>
-        <script type="module" src="./asset/js/home.js"></script>
+        <script type="module" src="asset/js/dividend.js"></script>
     </body>
 </html>
